backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Data Processing
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.13.0
torch==2.1.0
torchvision==0.16.0

# Financial Data
yfinance==0.2.28
alpha-vantage==2.3.1
ta==0.10.2

# Reinforcement Learning
gymnasium==0.29.1
stable-baselines3==2.2.1

# Database - MySQL
mysqlclient==2.2.0
sqlalchemy==2.0.23
alembic==1.12.1

# Cache & Sessions
redis==5.0.1
celery==5.3.4

# Networking
websockets==12.0
requests==2.31.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Financial specific
backtrader==1.9.78.123
zipline-reloaded==3.0.4
vectorbt==0.25.2 